# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Readora API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /genres:
    get:
      operationId: listGenres
      tags: [genres]
      summary: List all genres
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
    post:
      operationId: createGenre
      tags: [genres]
      summary: Create genre (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"
  /genres/{id}:
    patch:
      operationId: updateGenre
      tags: [genres]
      summary: Update genre (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"
    delete:
      operationId: deleteGenre
      tags: [genres]
      summary: Delete genre (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CYCLES
...
  /books/{id}/chapters:
    get:
      operationId: listChapters
      tags: [reader]
      summary: Get book table of contents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /books/{id}/chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [reader]
      summary: Get chapter content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterContent"
  /books/{id}/progress:
    get:
      operationId: getProgress
      tags: [reader]
      summary: Get reading progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingProgress"
    put:
      operationId: saveProgress
      tags: [reader]
      summary: Save reading progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingProgress"
  /reader/settings:
    get:
      operationId: getReaderSettings
      tags: [reader]
      summary: Get reader settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderSettings"
    put:
      operationId: saveReaderSettings
      tags: [reader]
      summary: Save reader settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderSettingsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderSettings"
    delete:
      operationId: resetReaderSettings
      tags: [reader]
      summary: Reset reader settings to defaults
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderSettings"

  # PUBLIC
...
components:
  schemas:
...
    Genre:
      type: object
      required: [id, code, name, isActive]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
    GenreInput:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
    GenreUpdate:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean

    # CYCLES
...
    Chapter:
      type: object
      required: [id, bookId, index, title]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        index:
          type: integer
        title:
          type: string
        wordCount:
          type: ["integer", "null"]
    ChapterContent:
      type: object
      required: [id, bookId, index, title, htmlContent]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        index:
          type: integer
        title:
          type: string
        htmlContent:
          type: string
        wordCount:
          type: ["integer", "null"]
    ReadingProgress:
      type: object
      required: [bookId, userId]
      properties:
        bookId:
          type: integer
        userId:
          type: integer
        currentChapterId:
          type: ["integer", "null"]
        currentPosition:
          type: ["number", "null"]
        progressPercent:
          type: ["number", "null"]
        readingStatus:
          type: string
          enum: [not_started, reading, finished]
        lastReadAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
    ProgressInput:
      type: object
      properties:
        currentChapterId:
          type: integer
        currentPosition:
          type: number
        progressPercent:
          type: number
        readingStatus:
          type: string
          enum: [not_started, reading, finished]
    ReaderSettings:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
        fontSize:
          type: integer
        fontFamily:
          type: string
        lineHeight:
          type: number
        theme:
          type: string
          enum: [light, dark, sepia]
        contentWidth:
          type: integer
    ReaderSettingsInput:
      type: object
      properties:
        fontSize:
          type: integer
        fontFamily:
          type: string
        lineHeight:
          type: number
        theme:
          type: string
          enum: [light, dark, sepia]
        contentWidth:
          type: integer

    # ADMIN
...